Move the `patchworkpp.h` file to `patchwork/patchworkpp.h` for installation in the local environment.
